update:除时间戳以外，其他整数变量都用int

+++ mojiayi-the-rich/dao/Models.go
@@ -11,23 +11,23 @@ import (
 )
 
 type BaseModel struct {
-	ID         uint64    `gorm:"primary_key" json:"id"`
+	ID         uint      `gorm:"primary_key" json:"id"`
 	CreateTime time.Time `json:"create_time"`
 	UpdateTime time.Time `json:"update_time"`
 	DeleteFlag uint8     `json:"delete_flag"`
 }
 
 type BasePageResult struct {
-	CurrentPage int32       `json:"currentPage"`
-	PageSize    int32       `json:"pageSize"`
-	Total       int32       `json:"total"`
-	Pages       int32       `json:"pages"`
+	CurrentPage int         `json:"currentPage"`
+	PageSize    int         `json:"pageSize"`
+	Total       int         `json:"total"`
+	Pages       int         `json:"pages"`
 	Data        interface{} `json:"data"`
 }
 
 type PageInfo struct {
-	CurrentPage int32 `json:"currentPage"`
-	PageSize    int32 `json:"pageSize"`
+	CurrentPage int `json:"currentPage"`
+	PageSize    int `json:"pageSize"`
 }
 
 var DB *gorm.DB
